feat: modified function for module
