+++ client/src/js/main.js
@@ -4,3 +4,4 @@ import menuController from './controllers/menuController';
 import menuModalController from './controllers/menuModalController';
 import navController from './controllers/navController';
 import footerController from './controllers/footerController';
+import refreshPage from './controllers/refreshPage';
