+++ client/src/sass/partials/_menu-section.scss
@@ -100,7 +100,6 @@
 
   .modal {
     width: 500px;
-    height: 420px;
     background: #fff;
     border-radius: 10px;
